Add virt-template

+++ renovate.sh
@@ -6,10 +6,34 @@ cd /home/dholler/repos/github.com/dominikholler/kubevirt-renovate
 GITHUB_PAT=
 GITHUB_AUTHOR_TOKEN=
 LOG_LEVEL=INFO
+#RENOVATE_DRY_RUN=full
+RENOVATE_DRY_RUN=null
+
+UPDATE_TEMPLATE=true
+UPDATE_SSP=false
+UPDATE_KUBEVIRT=false
 
 
-buildah build --pull=always  -f Containerfile.ssp   -t kubevirt-renovate-ssp
-podman run --rm -it \
+if $UPDATE_TEMPLATE ; then
+  REPO=kubevirt/virt-template
+  #REPO=tmp-kv-mirror/virt-template
+
+  podman run --rm -it --pull=always \
+    -e RENOVATE_FORK_TOKEN=$GITHUB_PAT \
+    -e RENOVATE_TOKEN=$GITHUB_AUTHOR_TOKEN \
+    -e RENOVATE_REPOSITORIES=$REPO \
+    -e LOG_LEVEL=$LOG_LEVEL  \
+    -e RENOVATE_CONFIG="$(< virt-template-renovate.json)" \
+    -e RENOVATE_ALLOWED_POST_UPGRADE_COMMANDS='["make vendor", "make all"]' \
+    -e RENOVATE_ONBOARDING=false  \
+    -e RENOVATE_DRY_RUN=$RENOVATE_DRY_RUN \
+    ghcr.io/renovatebot/renovate
+
+fi
+
+if $UPDATE_SSP ; then
+  buildah build --pull=always  -f Containerfile.ssp   -t kubevirt-renovate-ssp
+  podman run --rm -it \
     -e RENOVATE_FORK_TOKEN=$GITHUB_PAT \
     -e RENOVATE_TOKEN=$GITHUB_AUTHOR_TOKEN \
     -e RENOVATE_REPOSITORIES=kubevirt/ssp-operator \
@@ -18,15 +42,14 @@ podman run --rm -it \
     -e RENOVATE_CONFIG="$(< ssp-renovate.json)" \
     -e RENOVATE_ONBOARDING=false  \
     localhost/kubevirt-renovate-ssp
+fi
 
-
-
-buildah build --pull=always --build-arg BAZEL_VERSION=6.5.0 --format docker -t kubevirt-renovate-bazel-650
-
+if $UPDATE_KUBEVIRT ; then
 REPO=kubevirt/kubevirt
 #REPO=tmp-kv-mirror/kubevirttest
 
-podman run --rm -it \
+  buildah build --pull=always --build-arg BAZEL_VERSION=6.5.0 --format docker -t kubevirt-renovate-bazel-650
+  podman run --rm -it \
     -e RENOVATE_FORK_TOKEN=$GITHUB_PAT \
     -e RENOVATE_TOKEN=$GITHUB_AUTHOR_TOKEN \
     -e RENOVATE_REPOSITORIES=$REPO \
@@ -40,8 +63,8 @@ podman run --rm -it \
     localhost/kubevirt-renovate-bazel-650
 
 
-buildah build --pull=always --build-arg BAZEL_VERSION=5.4.1 --format docker -t kubevirt-renovate-bazel-541
-podman run --rm -it \
+  buildah build --pull=always --build-arg BAZEL_VERSION=5.4.1 --format docker -t kubevirt-renovate-bazel-541
+  podman run --rm -it \
     -e RENOVATE_FORK_TOKEN=$GITHUB_PAT \
     -e RENOVATE_TOKEN=$GITHUB_AUTHOR_TOKEN \
     -e RENOVATE_REPOSITORIES=$REPO \
@@ -54,7 +77,7 @@ podman run --rm -it \
     -e RENOVATE_BASE_BRANCH_PATTERNS='["release-1.6"]' \
     localhost/kubevirt-renovate-bazel-541
 
-podman run --rm -it \
+  podman run --rm -it \
     -e RENOVATE_FORK_TOKEN=$GITHUB_PAT \
     -e RENOVATE_TOKEN=$GITHUB_AUTHOR_TOKEN \
     -e RENOVATE_REPOSITORIES=$REPO \
@@ -65,9 +88,5 @@ podman run --rm -it \
     -e RENOVATE_REQUIRE_CONFIG=optional \
     -e RENOVATE_DRY_RUN=$RENOVATE_DRY_RUN \
     localhost/kubevirt-renovate-bazel-541
-
-
-
-
-
+fi
